requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
python-telegram-bot==20.7
Werkzeug==3.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0